refactor: extract PDF continuation evidence
